+++ src/i18n/languages/csCZ.ts
@@ -225,6 +225,7 @@ const dictionary: RawTranslationPackage = {
     UNICHAR: 'UNICHAR',
     UNICODE: 'UNICODE',
     UPPER: 'VELKÁ',
+    VALUE: 'HODNOTA',
     VARA: 'VARA',
     'VAR.P': 'VAR.P',
     VARPA: 'VARPA',