+++ src/i18n/languages/trTR.ts
@@ -225,6 +225,7 @@ const dictionary: RawTranslationPackage = {
     UNICHAR: 'UNICODEKARAKTERİ',
     UNICODE: 'UNICODE',
     UPPER: 'BÜYÜKHARF',
+    VALUE: 'DEĞER',
     VARA: 'VARA',
     'VAR.P': 'VAR.P',
     VARPA: 'VARSA',