+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added a new function: IRR. [#1591](https://github.com/handsontable/hyperformula/issues/1591)
 - Added a new function: N. [#1585](https://github.com/handsontable/hyperformula/issues/1585)
+- Added a new function: VALUE. [#1592](https://github.com/handsontable/hyperformula/issues/1592)
 
 ## [3.1.1] - 2025-12-18